+++ db/migrate/20220717155034_add_disable_dataset_requests_to_dataset.rb
@@ -1,0 +1,6 @@
+class AddDisableDatasetRequestsToDataset < ActiveRecord::Migration[6.1]
+  def change
+    add_column :datasets, :disable_data_requests, :boolean, null: false, default: false
+    add_index :datasets, :disable_data_requests
+  end
+end